+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+* Can handle leading and trailing dashes
+
 # 0.0.1
 
 Initial version with basic functionality